remove workaround for compiler bug
