+++ src/lib.rs
@@ -402,9 +402,113 @@ pub mod unsync {
 
 #[macro_use]
 pub mod sync {
-    // Can't use `OnceCell(imp::OnceCell) due to
-    // https://github.com/rust-lang/rust/issues/50518
-    pub use imp::OnceCell;
+    use imp::OnceCell as Imp;
+
+    /// A thread-safe cell which can be written to only once.
+    ///
+    /// Unlike `::std::sync::Mutex`, a `OnceCell` provides simple `&`
+    /// references to the contents.
+    ///
+    /// # Example
+    /// ```
+    /// use once_cell::sync::OnceCell;
+    ///
+    /// static CELL: OnceCell<String> = OnceCell::INIT;
+    /// assert!(CELL.get().is_none());
+    ///
+    /// ::std::thread::spawn(|| {
+    ///     let value: &String = CELL.get_or_init(|| {
+    ///         "Hello, World!".to_string()
+    ///     });
+    ///     assert_eq!(value, "Hello, World!");
+    /// }).join().unwrap();
+    ///
+    /// let value: Option<&String> = CELL.get();
+    /// assert!(value.is_some());
+    /// assert_eq!(value.unwrap().as_str(), "Hello, World!");
+    /// ```
+    #[derive(Debug)]
+    pub struct OnceCell<T>(Imp<T>);
+
+    impl<T> Default for OnceCell<T> {
+        fn default() -> OnceCell<T> {
+            OnceCell::new()
+        }
+    }
+
+    impl<T> From<T> for OnceCell<T> {
+        fn from(value: T) -> Self {
+            let cell = Self::new();
+            cell.get_or_init(|| value);
+            cell
+        }
+    }
+
+    impl<T: PartialEq> PartialEq for OnceCell<T> {
+        fn eq(&self, other: &OnceCell<T>) -> bool {
+            self.get() == other.get()
+        }
+    }
+
+    impl<T> OnceCell<T> {
+        /// An empty cell, for initialization in a `const` context.
+        pub const INIT: OnceCell<T> = OnceCell(Imp::INIT);
+
+        /// Creates a new empty cell.
+        pub fn new() -> OnceCell<T> {
+            OnceCell(Imp::new())
+        }
+
+        /// Gets the reference to the underlying value. Returns `None`
+        /// if the cell is empty.
+        pub fn get(&self) -> Option<&T> {
+            self.0.get()
+        }
+
+        /// Sets the contents of this cell to `value`. Returns
+        /// `Ok(())` if the cell was empty and `Err(value)` if it was
+        /// full.
+        ///
+        /// # Example
+        /// ```
+        /// use once_cell::sync::OnceCell;
+        ///
+        /// static CELL: OnceCell<i32> = OnceCell::INIT;
+        ///
+        /// fn main() {
+        ///     assert!(CELL.get().is_none());
+        ///
+        ///     ::std::thread::spawn(|| {
+        ///         assert_eq!(CELL.set(92), Ok(()));
+        ///     }).join().unwrap();
+        ///
+        ///     assert_eq!(CELL.set(62), Err(62));
+        ///     assert_eq!(CELL.get(), Some(&92));
+        /// }
+        /// ```
+        pub fn set(&self, value: T) -> Result<(), T> {
+            self.0.set(value)
+        }
+
+        /// Gets the contents of the cell, initializing it with `f`
+        /// if the cell was empty. May threads may call `get_or_init`
+        /// concurrently with different initializing functions, but
+        /// it is guaranteed that only one function will be executed.
+        ///
+        /// # Example
+        /// ```
+        /// use once_cell::sync::OnceCell;
+        ///
+        /// let cell = OnceCell::new();
+        /// let value = cell.get_or_init(|| 92);
+        /// assert_eq!(value, &92);
+        /// let value = cell.get_or_init(|| unreachable!());
+        /// assert_eq!(value, &92);
+        /// ```
+        pub fn get_or_init<F: FnOnce() -> T>(&self, f: F) -> &T {
+            self.0.get_or_init(f)
+        }
+    }
 
     /// A value which is initialized on the first access.
     ///
